requirements.txt
dj-database-url==0.5.0
django==3.0.4
psycopg2-binary==2.8.4
pytest-cov==2.8.1
pytest-django==3.8.0
